feat(task): implement new task submission/verification flow
The v2.0 implementation of the task submission and verification flow provides a cleaner and more discreet user experience. Each task event contains a Submit Screenshot button, which allows the user to privately submit their task via bot DM for approval where it will then be sent to the task admins in a private channel to approve or reject (with optional reason given).