feat(movie): implement pickmovie command for movie night
The /pickmovie command is a complicated beast of a flow, presenting the admin with four different options to select a movie for movie night. The first is to search for a specific movie using a modal and Fuse fuzzy matching logic. The second is to roll a random movie, allowing for re-rolls before confirmation. The third is a public poll with preview embeds for each movie, randomly selected with only one option per "added by" user. The fourth is a poll built from specific selections by the admin, using a paginated string selector dropdown to build the poll with.