feat(movie): implement movie lifecycle handling
The movie lifecycle handles the ending of movie nights either via the admin-only /endmovie command or automatically calculated based on the movie's runtime + buffer, removing the watched movie from the list and adding it to a stored history file. This commit also includes some improvements to the reminders and overall functionality.