chore(task): create scripts for legacy task completion data import
Added a script to import the v1 task submission count for each user to Firestore into the respective userStats document as a "legacyTasksCompleted" value. Additionally added a block of code into the handlePrizeDraw to import the final legacy completions into the first prize draw in v2. This block will be removed in the 2.1 update.