fix(core): implement minor bug fixes
Fixed some minor bugs in src and tests: changed the formatting of the "Added By" field on embeds to use * for italics rather than _ due to the presence of underscores in some usernames. Tests were updated to handle minor intentional source code changes to prevent false negatives in test runs.